demo/producer/requirements.txt
nats-py>=2.7.0
jsonschema>=4.21.0
requests>=2.31.0
